consumer/requirements.txt
kafka-python
numpy
opencv-python
